% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussl0.R
\name{gaussl0}
\alias{gaussl0}
\title{fit a linear regression model with l0 penalty}
\usage{
gaussl0(X, y, lam, maxit, eps)
}
\arguments{
\item{X}{Input matrix, of dimension nobs x nvars; each row is an observation vector}

\item{y}{Response variable(quantitative variable)}

\item{lam}{A user supplied \code{lambda} value}

\item{maxit}{Maximum number of passes over the data for \code{lambda}}

\item{eps}{Convergence threshold}
}
\value{
\item{beta}{a vector of coefficients}
 \item{iter}{number of iterations}
}
\description{
l0 iterative regression
}
\author{
Wenchuan Guo <wguo007@ucr.edu>
}

