% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot.cv.l0ara}
\alias{plot.cv.l0ara}
\title{plot for an "cv.l0ara" object}
\usage{
\method{plot}{cv.l0ara}(x, col = 3, ...)
}
\arguments{
\item{x}{Fitted "cv.l0ara" object.}

\item{col}{color of the dots.}

\item{...}{Not used argument.}
}
\description{
Produces curves from a fitted "cv.l0ara" object.
}
\seealso{
\code{\link{predict}}, \code{\link{coef}} methods, \code{\link{cv.l0ara}} and \code{\link{l0ara}} function.
}
\author{
Wenchuan Guo <wguo007@ucr.edu>, Shujie Ma <shujie.ma@ucr.edu>, Zhenqiu Liu <Zhenqiu.Liu@cshs.org>
}
