% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot.l0ara}
\alias{plot.l0ara}
\title{plot for a "l0ara" object}
\usage{
\method{plot}{l0ara}(x, auc = FALSE, ...)
}
\arguments{
\item{x}{Fitted "l0ara" object.}

\item{auc}{logical; if \code{TRUE}, produces \code{auc} curve for "family="logit""}

\item{...}{Not used argument.}
}
\description{
Auc plot for logsitic regression
}
\details{
Wenchuan Guo <wguo007@ucr.edu>, Zhenqiu Liu <Zhenqiu.Liu@cshs.org>
}
\seealso{
\code{predict}, \code{coef} methods and \code{l0ara} function.
}

