% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{predict.l0ara}
\alias{predict.l0ara}
\title{make predictions from a "l0ara" object.}
\usage{
\method{predict}{l0ara}(object, newx, type = c("link", "response",
  "coefficients", "class"), ...)
}
\arguments{
\item{object}{Fitted "l0ara" object.}

\item{newx}{Matrix of new values for x at which predictions are to be made. Must be a matrix.}

\item{type}{Type of prediction required. "link" gives the linear predictors(for "gaussian" models it gives the fitted values). "response" gives the fitted probabilities for "logit" and fitted mean for "poisson". "coefficients" gives the coefficients which is same as "coef" function. "class" (applies only to "logit") produces the class label corresponding to the maximum probability.}

\item{...}{Not used argument.}
}
\value{
The object returned depends the functions.
}
\description{
Make predictions from the model.
}
\details{
This function makes it easier to use the results to make a prediction or to see the fitted model.
}
\seealso{
\code{\link{coef}} method and \code{\link{l0ara}} function.
}
\author{
Wenchuan Guo <wguo007@ucr.edu>, Shujie Ma <shujie.ma@ucr.edu>, Zhenqiu Liu <Zhenqiu.Liu@cshs.org>
}
