% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.cv.l0ara}
\alias{print.cv.l0ara}
\title{summarizing the fits from a "cv.l0ara" object.}
\usage{
\method{print}{cv.l0ara}(x, ...)
}
\arguments{
\item{x}{Fitted "cv.l0ara" object.}

\item{...}{Not used argument.}
}
\description{
Print the general information of the cross validated fit.
}
\details{
This function makes it easier to see the cross-validation results.
}
\seealso{
\code{\link{predict}}, \code{\link{coef}} methods and \code{\link{l0ara}} function.
}
\author{
Wenchuan Guo <wguo007@ucr.edu>, Shujie Ma <shujie.ma@ucr.edu>, Zhenqiu Liu <Zhenqiu.Liu@cshs.org>
}
