% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{coef.l0ara}
\alias{coef.l0ara}
\title{print coefficients from a "l0ara" object.}
\usage{
\method{coef}{l0ara}(object, ...)
}
\arguments{
\item{object}{Fitted "l0ara" object.}

\item{...}{Not used argument.}
}
\value{
The object returns the coefficients.
}
\description{
Print the coefficients from the model.
}
\details{
This function makes it easier to use the results to make a prediction or to see the fitted model.
}
\author{
Wenchuan Guo <wguo007@ucr.edu>
}
\seealso{
\code{predict} method and \code{l0ara} function.
}

