% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.l0ara.R
\name{cv.l0ara}
\alias{cv.l0ara}
\title{Cross validation for l0ara}
\usage{
cv.l0ara(x, y, family, lam, measure, nfolds, maxit, eps, seed)
}
\arguments{
\item{x}{Input matrix as in \code{l0ara}}

\item{y}{Response variable as in \code{l0ara}}

\item{family}{Response type as in \code{l0ara}}

\item{lam}{A user supplied \code{lambda} sequence in descending or asecending order.}

\item{measure}{Loss function used for corss validation. \code{measurer="mse"} or \code{"mae"} for all models. \code{"measure"="class"} only for logsitic regression.}

\item{nfolds}{Number of folds. Default value is 10. Smallest value is 3.}

\item{maxit}{Maximum number of passes over the data for \code{lambda}}

\item{eps}{Convergence threshold}

\item{seed}{Seed of random number generator}
}
\value{
\item{cv.error}{The mean cross validated error for given lambda sequence}
 \item{lam.min}{The lambda gives min cv.error}
 \item{lambda}{The lambda used}
 \item{measure}{type of measure}
 \item{family}{model used}
Wenchuan Guo <wguo007@ucr.edu>, Zhenqiu Liu <Zhenqiu.Liu@cshs.org>
}
\description{
Does k-fold cross-validation for l0ara, produces a plot, and returns the optimal \code{lambda}
}
\examples{
#' # Linear regression
# Generate design matrix and response variable
n <- 100
p <- 40
x <- matrix(rnorm(n*p), n, p)
beta <- c(1,0,2,3,rep(0,p-4))
noise <- rnorm(n)
y <- x\%*\%beta+noise
lam <- c(0.1, 0.3, 0.5)
fit <- cv.l0ara(x, y, family="gaussian", lam, measure = "mse")
}
\seealso{
\code{l0ara}.
}

