% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0ara.R
\name{l0ara}
\alias{l0ara}
\title{fit a generalized linear model with l0 penalty}
\usage{
l0ara(x, y, family, lam, maxit, eps)
}
\arguments{
\item{x}{Input matrix, of dimension nobs x nvars; each row is an observation vector}

\item{y}{Response variable. Quantitative for \code{family="gaussian"}; positive quantitative for \code{family="gamma"} or \code{family="inv.gaussian"} ; a factor with two levels for \code{family="logit"}; non-negative counts for \code{family="poisson"}}

\item{family}{Response type(see above)}

\item{lam}{A user supplied \code{lambda} value}

\item{maxit}{Maximum number of passes over the data for \code{lambda}}

\item{eps}{Convergence threshold}
}
\value{
\item{beta}{a vector of coefficients}
 \item{df}{number of nonzero coefficients}
 \item{iter}{number of iterations}
 \item{lambda}{The lambda used}
}
\description{
An efficient algorithm for feature selection with L0 penalty.
}
\examples{
# Linear regression
# Generate design matrix and response variable
n <- 100
p <- 40
x <- matrix(rnorm(n*p), n, p)
beta <- c(1,0,2,3,rep(0,p-4))
noise <- rnorm(n)
y <- x\%*\%beta+noise
# fit sparse linear regression
res.gaussian <- l0ara(x, y, family="gaussian", 1)

# predict for new observations
print(res.gaussian)
predict(res.gaussian, newx=matrix(rnorm(3,p),3,p))
coef(res.gaussian)

# Logistic regression
# Generate design matrix and response variable
n <- 100
p <- 40
x <- matrix(rnorm(n*p), n, p)
beta <- c(1,0,2,3,rep(0,p-4))
prob <- exp(x\%*\%beta)/(1+exp(x\%*\%beta))
y <- rbinom(n, rep(1,n), prob)
# fit sparse logistic regression
res.logit <- l0ara(x, y, family="logit", 0.7)

# predict for new observations
print(res.logit)
predict(res.logit, newx=matrix(rnorm(3,p),3,p))
coef(res.logit)

# Poisson regression
# Generate design matrix and response variable
n <- 100
p <- 40
x <- matrix(rnorm(n*p), n, p)
beta <- c(1,0,0.5,0.3,rep(0,p-4))
mu <- exp(x\%*\%beta)
y <- rpois(n, mu)
# fit sparse logistic regression
res.pois <- l0ara(x, y, family="pois", 0.7)

# predict for new observations
print(res.pois)
predict(res.pois, newx=matrix(rnorm(3,p),3,p))
coef(res.pois)
}
\author{
Wenchuan Guo <wguo007@ucr.edu>
}
\seealso{
\code{predict}, \code{coef} methods.
}

