% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot.l0ara}
\alias{plot.l0ara}
\title{plot for an "l0ara" object}
\usage{
\method{plot}{l0ara}(x, auc = FALSE, split = FALSE, col = 4, ...)
}
\arguments{
\item{x}{Fitted "l0ara" object.}

\item{auc}{logical; if \code{TRUE}, produces \code{auc} curve for \code{family=logit}.}

\item{split}{logical; if if \code{TRUE}, produces seperate plots.}

\item{col}{color of the dots.}

\item{...}{Not used argument.}
}
\description{
Two plots are availiable: a plot of fitted value against linear predictor; \code{roc}(\code{auc}) curve for \code{family="logit"}.
}
\seealso{
\code{\link{predict}}, \code{\link{coef}} methods and \code{\link{l0ara}} function.
}
\author{
Wenchuan Guo <wguo007@ucr.edu>, Shujie Ma <shujie.ma@ucr.edu>, Zhenqiu Liu <Zhenqiu.Liu@cshs.org>
}
