% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.l0ara}
\alias{print.l0ara}
\title{summarizing the fits from a "l0ara" object.}
\usage{
\method{print}{l0ara}(x, ...)
}
\arguments{
\item{x}{Fitted "l0ara" object.}

\item{...}{Not used argument.}
}
\description{
This function print the general information of the fit.
}
\details{
This function makes it easier to see the fitted model.
}
\author{
Wenchuan Guo <wguo007@ucr.edu>
}
\seealso{
\code{predict}, \code{coef} methods and \code{l0ara} function.
}

